requirements.txt
python-dotenv
editdistance
requests